requirements.txt
functions-framework>=3.0.0
google-cloud-storage>=2.0.0
google-cloud-secret-manager>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-mock>=3.0.0
requests>=2.28.0